Fix bug that caused too many listeners warning.
On longer playback or sync jobs the EventEmitter implementation
generated a warning message about a possible memory leak. The reason
for this was that the pipe() method adds listeners not only to the
readable side of the stream, but to the writable as well. So in this
fix we unpipe the converter from the stream after the playback is done
to remove the event listeners again.