Remove i18next dependency.
Calling `i18next.init` multiple times does create a new `t` function
but also removes the previously installed translations for previously
created `t` functions.

That does not seem to be a problem in a Node environment where the
resol-vbus module gets its own copy of the i18next library. But initial
tests in the browser caused the `I18N` module to fail because of the
shared use of the i18next module.

This commit removes the dependency to i18next and implements its own
restricted version of the `t` function.